Create reply-to-stackatron
OP tweet:

"Introducing Zin, a browser for the self-contained, immutable, Bitcoin Internet.

On Zin, web pages are recursive inscriptions, and URLs are SNS names.

A new internet, where all data lives on Bitcoin, is ready to be realized."

The Zin branding, as well as the design motif on one of the proof-of-concept sats-sites as shown here: https://zin.so/hellofelipe.sats

I noticed Bob Bodily commented on the Tweet. Bob is a leading ICP dev and personality. 

Zenon should make an impression, there.