Create reply-to-stackatron
Create reply-to-stackatron
This original Tweet is gaining momentum. As predicted, Ordinals are educating the crypto market about Bitcoin ethos. 

The reply content proposed herein is intended to lead the discussion toward the next level of decentralized internet: decentralized infrastructure. Community members may follow on and engage this reply to further the discussion.

Original tweet:
=====
"Introducing Zin, a browser for the self-contained, immutable, Bitcoin Internet.

On Zin, web pages are recursive inscriptions, and URLs are SNS names.

A new internet, where all data lives on Bitcoin, is ready to be realized."

The Zin branding, as well as the design motif on one of the proof-of-concept sats-sites as shown here: https://zin.so/hellofelipe.sats
======

With this reply, Zenon will make first contact and may be followed by a series of community alien encounters.